gracefuly handle missing text file
Prior to this change, the IOError on a missing text file would cause
Helga to fail to load the entrypoint on start. Additionally Helga would
log this entrypoint error on every single IRC action.

Gracefully handle the missing text file so we only see a single error in
the helga log at startup.